Issue29: Moved Model.js under framework folder and updated the imports accordingly
