+++ src/app/models/Crop.js
@@ -1,7 +1,7 @@
 const { DataTypes } = require('sequelize');
 const { Family } = require('./Family');
 const { Group } = require('./Group');
-const { Model } = require('./Model');
+const { Model } = require('../../framework/models/Model');
 
 /**
  * For more information on sequelize attributes & options
