+++ src/app/models/Observer.js
@@ -1,6 +1,6 @@
 const { DataTypes } = require('sequelize');
 const { Log } = require('../providers/LoggingProvider');
-const { Model } = require('./Model')
+const { Model } = require('../../framework/models/Model')
 
 /**
  * For more information on sequelize attributes & options
